% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backgroundSubtract.R
\name{BackgroundSubtract}
\alias{BackgroundSubtract}
\title{Pull out specific data from flowjo workspace}
\usage{
BackgroundSubtract(colData = colData, ChannelValueList = insolubleG1)
}
\arguments{
\item{colData}{a file that contains all your sample names and all associate labels}

\item{ChannelValueList}{this is the parsed workspace with gate info. This is essentially the output from getFlowjoWspData function}
}
\value{
A list of data frames with all the data specific to each gate. Each item in the list is a different samlpe and each data frame has the raw data.
}
\description{
Given a flowjo workspace and a list of .fcs files this code will generate a list of dataframes and specific gates can be specified.
}
