% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backgroundSubtractMedian.R
\name{BackgroundSubtract}
\alias{BackgroundSubtract}
\title{Background subtract the median control from y-axis column}
\usage{
BackgroundSubtract(colData = "colData", ChannelValueList = "ChannelValueList")
}
\arguments{
\item{colData}{a file that contains all your sample names and all associate labels}

\item{ChannelValueList}{this is the parsed workspace with gate info. This is essentially the output from getFlowjoWspData function}
}
\value{
A data frame with a column with background subtracted values. This output can be plotted with ggplot2
}
\description{
Given a parsed flowjo workspace and a colData file, this function will background subtract the median value of the control from the treated samples.
}
