% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFlowjoWspData.R
\name{getFlowjoWspData}
\alias{getFlowjoWspData}
\title{Pull out specific data from flowjo workspace}
\usage{
getFlowjoWspData(
  wspFilename = "04-Oct-2021.wsp",
  nodes = "Cells",
  markers = c("FSC-A", "SSC-A", "FL1-A")
)
}
\arguments{
\item{wspFilename}{wsp file that includes gating setting and other data from flowjo}

\item{nodes}{nodes specifies the gate being used for the analysis}

\item{markers}{marker specifies the columns of data you want back. Ex: FSC-A, SSC-A, FL1-A}
}
\value{
A list of data frames with all the data specific to each gate you specify. Each item in the list is a different sample and each data frame has the raw data.
}
\description{
Given a flowjo workspace that is in the same directory as the .fcs files, this code will generate a list of dataframes for the specific gates that are specified.
}
